Migration

+++ dose-em-dia-backend/DoseEmDia/DoseEmDia/Models/Localizacao/Cidade.cs
@@ -6,7 +6,7 @@ namespace DoseEmDia.Models.Localizacao
     {
         public long IdCidade { get; set; }
         public string Nome { get; set; } = default!;
-        public long EstadoId { get; set; }
+        public int EstadoId { get; set; }
         public Estado Estado { get; set; } = default!;
 
         public ICollection<Cep> Ceps { get; set; } = new List<Cep>();
