+++ dose-em-dia-backend/DoseEmDia/DoseEmDia/Models/Localizacao/Estado.cs
@@ -2,10 +2,10 @@
 {
     public class Estado
     {
-        public long IdEstado { get; set; }
+        public int IdEstado { get; set; }
         public string Nome { get; set; } = default!;
         public string Uf { get; set; } = default!;
-        public long PaisId { get; set; }
+        public int PaisId { get; set; }
         public Pais Pais { get; set; } = default!;
 
         public ICollection<Cidade> Cidades { get; set; } = new List<Cidade>();
